fix(ci): isolate landing release download tests
